@@ -1,12 +1,12 @@
 module k8gbterratest
 
-go 1.17
+go 1.19
 
 require (
 	github.com/AbsaOSS/gopkg v0.1.3
-	github.com/gruntwork-io/terratest v0.38.6
+	github.com/gruntwork-io/terratest v0.40.24
 	github.com/stretchr/testify v1.7.0
-	gopkg.in/yaml.v2 v2.4.0
+	gopkg.in/yaml.v3 v3.0.1
 	k8s.io/api v0.20.6
 	k8s.io/apimachinery v0.20.6
 )
@@ -30,7 +30,7 @@ require (
 	github.com/gruntwork-io/go-commons v0.8.0 // indirect
 	github.com/hashicorp/errwrap v1.0.0 // indirect
 	github.com/hashicorp/go-multierror v1.1.0 // indirect
-	github.com/imdario/mergo v0.3.5 // indirect
+	github.com/imdario/mergo v0.3.11 // indirect
 	github.com/jmespath/go-jmespath v0.4.0 // indirect
 	github.com/json-iterator/go v1.1.11 // indirect
 	github.com/lixiangzhong/dnsutil v0.0.0-20191203032812-75ad39d2945a // indirect
@@ -47,14 +47,14 @@ require (
 	golang.org/x/crypto v0.0.0-20210513164829-c07d793c2f9a // indirect
 	golang.org/x/net v0.0.0-20210614182718-04defd469f4e // indirect
 	golang.org/x/oauth2 v0.0.0-20210514164344-f6687ab2804c // indirect
-	golang.org/x/sys v0.0.0-20210603125802-9665404d3644 // indirect
+	golang.org/x/sys v0.0.0-20220517195934-5e4e11fc645e // indirect
 	golang.org/x/term v0.0.0-20201126162022-7de9c90e9dd1 // indirect
 	golang.org/x/text v0.3.6 // indirect
 	golang.org/x/time v0.0.0-20200630173020-3af7569d3a1e // indirect
 	google.golang.org/appengine v1.6.7 // indirect
 	google.golang.org/protobuf v1.26.0 // indirect
 	gopkg.in/inf.v0 v0.9.1 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
 	k8s.io/client-go v0.20.6 // indirect
 	k8s.io/klog/v2 v2.4.0 // indirect
 	k8s.io/utils v0.0.0-20201110183641-67b214c5f920 // indirect
@@ -62,26 +62,10 @@ require (
 	sigs.k8s.io/yaml v1.2.0 // indirect
 )
 
-replace (
-	github.com/Azure/go-autorest/autorest/adal => github.com/Azure/go-autorest/autorest/adal v0.9.16
-	github.com/containerd/containerd => github.com/containerd/containerd v1.5.8
-	github.com/gogo/protobuf v1.3.1 => github.com/gogo/protobuf v1.3.2
-	github.com/opencontainers/image-spec v1.0.1 => github.com/opencontainers/image-spec v1.0.2
-	github.com/opencontainers/runc v1.0.2 => github.com/opencontainers/runc v1.0.3
-	gopkg.in/yaml.v3 => gopkg.in/yaml.v3 v3.0.0
-)
+replace github.com/gogo/protobuf v1.3.1 => github.com/gogo/protobuf v1.3.2
 
 exclude (
-	github.com/Azure/go-autorest/autorest/adal v0.5.0
-	github.com/Azure/go-autorest/autorest/adal v0.8.2
 	github.com/Azure/go-autorest/autorest/adal v0.9.0
 	github.com/Azure/go-autorest/autorest/adal v0.9.2
 	github.com/Azure/go-autorest/autorest/adal v0.9.5
-	github.com/containerd/containerd v1.4.9
-	github.com/containerd/containerd v1.5.0-beta.1
-	github.com/containerd/containerd v1.5.0-beta.3
-	github.com/containerd/containerd v1.5.0-beta.4
-	github.com/containerd/containerd v1.5.0-rc.0
-	github.com/containerd/containerd v1.5.1
-	github.com/containerd/containerd v1.5.2
 )
