@@ -21,7 +21,6 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"reflect"
 	"strconv"
@@ -1547,7 +1546,7 @@ func configureEnvVar(config Config) {
 
 func getTestContext(testData string) (client.Client, *k8gbv1beta1.Gslb) {
 	// Create a fake client to mock API calls.
-	var gslbYaml, err = ioutil.ReadFile(testData)
+	var gslbYaml, err = os.ReadFile(testData)
 	if err != nil {
 		panic(fmt.Errorf("can't open example CR file: %s", testData))
 	}
