@@ -21,7 +21,6 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
 	"net"
 	"os"
 	"strings"
@@ -1280,7 +1279,7 @@ func provideSettings(t *testing.T, expected depresolver.Config) (settings testSe
 	if os.IsNotExist(err) {
 		t.Fatalf("Sample CR yaml file not found at: %s", crSampleYaml)
 	}
-	gslbYaml, err := ioutil.ReadFile(crSampleYaml)
+	gslbYaml, err := os.ReadFile(crSampleYaml)
 	if err != nil {
 		t.Fatalf("Can't open example CR file: %s", crSampleYaml)
 	}
