@@ -20,7 +20,7 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 
 import (
 	"fmt"
-	"io/ioutil"
+	"os"
 	"reflect"
 	"testing"
 
@@ -67,7 +67,7 @@ var a = struct {
 	},
 	Gslb: func() *k8gbv1beta1.Gslb {
 		var crSampleYaml = "../../../deploy/crds/k8gb.absa.oss_v1beta1_gslb_cr.yaml"
-		gslbYaml, _ := ioutil.ReadFile(crSampleYaml)
+		gslbYaml, _ := os.ReadFile(crSampleYaml)
 		gslb, _ := utils.YamlToGslb(gslbYaml)
 		return gslb
 	}(),
