@@ -21,8 +21,8 @@ Generated by GoLic, for more details see: https://github.com/AbsaOSS/golic
 
 import (
 	"fmt"
-	"io/ioutil"
 	"log"
+	"os"
 	"sort"
 	"strings"
 	"testing"
@@ -101,7 +101,7 @@ func DoWithRetryWaitingForValueE(t *testing.T, actionDescription string, maxRetr
 }
 
 func CreateGslb(t *testing.T, options *k8s.KubectlOptions, settings TestSettings, kubeResourcePath string) {
-	k8sManifestBytes, err := ioutil.ReadFile(kubeResourcePath)
+	k8sManifestBytes, err := os.ReadFile(kubeResourcePath)
 	if err != nil {
 		log.Fatal(err)
 	}
